Reduce hybrid search coupling by splitting execution paths
Hybrid search had grown into a high-connectivity file that made ranking, vector execution, and enrichment changes share one blast radius. Split the engine into focused modules while preserving the public hybrid-search-engine export surface.

Constraint: Issue #583 requires sub-500-line search modules, benchmark/search regression coverage, graphify refresh, and DEPRECATED cleanup.

Rejected: Keep private helpers inside HybridSearchEngine | would preserve the god-node shape and fail the module-boundary objective.

Confidence: high

Scope-risk: moderate

Directive: Keep hybrid-search-engine.ts as orchestration and compatibility exports; put new execution or ranking detail in focused sibling modules.

Tested: npm run type-check -w @memento/core; npm run build -w @memento/core; npx vitest run packages/memento-core/src/domains/search/algorithms/__tests__/hybrid-search-engine.spec.ts packages/memento-core/src/domains/search/algorithms/__tests__/hybrid-search-engine-consolidation.spec.ts packages/memento-core/src/domains/search/algorithms/__tests__/search-result-combiner-consolidation.spec.ts packages/memento-core/src/domains/search/algorithms/__tests__/search-ranking.spec.ts; npm run quality:benchmark:tune-weights -- --candidates 1 --seed 583 --output-dir /tmp/memento-tune-583; npm run lint:ts; python3 -c "from graphify.watch import _rebuild_code; from pathlib import Path; _rebuild_code(Path('.'))"; git diff --check

Not-tested: Full repository test suite.
